PAC4Ex3/src/main/java/edu/uoc/pac4/Merchandising.java
package edu.uoc.pac4;

import edu.uoc.pac4.exception.ProductException;

public class Merchandising extends Product implements Manufacturable{
    private double fabricationCost;
    private double packagingCost;
    private static final String DESCRIPTION = "A merchandising item";

    public Merchandising(String name, double price, double fabricationCost, double packagingCost) throws ProductException {
        super(name, price);
        setFabricationCost(fabricationCost);
        setPackagingCost(packagingCost);
    }

    public double getFabricationCost() {
        return fabricationCost;
    }

    public void setFabricationCost(double fabricationCost) {
        this.fabricationCost = fabricationCost;
    }

    public double getPackagingCost() {
        return packagingCost;
    }

    public void setPackagingCost(double packagingCost) {
        this.packagingCost = packagingCost;
    }

    @Override
    public double auditBenefits() {
        return (getPrice()-(getFabricationCost()+getPackagingCost()))*getSoldUnits();
    }

    @Override
    public String describeProduct() {
        return this.getClass().getSimpleName() + " (MANUFACTURED): " + DESCRIPTION;
    }
}
